# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProposalStatus =
  (typeof ProposalStatus)[keyof typeof ProposalStatus];

export const ProposalStatus = {
  draft: "draft",
  exported: "exported",
} as const;

export interface Proposal {
  id: number;
  clientName: string;
  industry: string;
  status: ProposalStatus;
  briefText: string;
  proposalContent: string;
  googleDocUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProposalBody {
  clientName: string;
  industry: string;
  briefText: string;
  proposalContent: string;
}

export type UpdateProposalBodyStatus =
  (typeof UpdateProposalBodyStatus)[keyof typeof UpdateProposalBodyStatus];

export const UpdateProposalBodyStatus = {
  draft: "draft",
  exported: "exported",
} as const;

export interface UpdateProposalBody {
  clientName?: string;
  industry?: string;
  proposalContent?: string;
  status?: UpdateProposalBodyStatus;
}

export interface ParseBriefBody {
  briefText: string;
}

export interface ParseBriefResponse {
  clientName: string;
  industry: string;
  proposalContent: string;
}

export interface GoogleDocsExportResponse {
  docId: string;
  docUrl: string;
  title: string;
}

export interface Tender {
  id: number;
  title: string;
  agency: string;
  description: string;
  category: string;
  deadline?: string | null;
  valueAmount?: string | null;
  sourceUrl?: string | null;
  contactInfo?: string | null;
  status: string;
  recommendationScore: number;
  proposalId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTenderBody {
  title: string;
  agency: string;
  description: string;
  category?: string;
  deadline?: string | null;
  valueAmount?: string | null;
  sourceUrl?: string | null;
  contactInfo?: string | null;
}

export interface ExtractTenderFromTextBody {
  text: string;
  sourceUrl?: string | null;
}

export interface ImportTendersCsvBody {
  csv: string;
}

export interface ImportTendersCsvResponse {
  imported: number;
  skipped: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListTendersParams = {
  /**
   * If true, only return tenders with recommendationScore > 0
   */
  recommended?: boolean;
};