# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /proposals:
    get:
      operationId: listProposals
      tags: [proposals]
      summary: List all proposals
      responses:
        "200":
          description: List of proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"
    post:
      operationId: createProposal
      tags: [proposals]
      summary: Create a new proposal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProposalBody"
      responses:
        "201":
          description: Created proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}:
    get:
      operationId: getProposal
      tags: [proposals]
      summary: Get a proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProposal
      tags: [proposals]
      summary: Update a proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProposalBody"
      responses:
        "200":
          description: Updated proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProposal
      tags: [proposals]
      summary: Delete a proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /proposals/parse-brief:
    post:
      operationId: parseBrief
      tags: [proposals]
      summary: Parse a project brief and generate a proposal using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParseBriefBody"
      responses:
        "200":
          description: Generated proposal content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParseBriefResponse"
...
  /proposals/{id}/export-to-google-docs:
    post:
      operationId: exportToGoogleDocs
      tags: [proposals]
      summary: Export a proposal to Google Docs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Google Docs export result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoogleDocsExportResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    Proposal:
      type: object
      properties:
        id:
          type: integer
        clientName:
          type: string
        industry:
          type: string
        status:
          type: string
          enum: [draft, exported]
        briefText:
          type: string
        proposalContent:
          type: string
        googleDocUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clientName
        - industry
        - status
        - briefText
        - proposalContent
        - createdAt
        - updatedAt
    CreateProposalBody:
      type: object
      properties:
        clientName:
          type: string
        industry:
          type: string
        briefText:
          type: string
        proposalContent:
          type: string
      required:
        - clientName
        - industry
        - briefText
        - proposalContent
    UpdateProposalBody:
      type: object
      properties:
        clientName:
          type: string
        industry:
          type: string
        proposalContent:
          type: string
        status:
          type: string
          enum: [draft, exported]
    ParseBriefBody:
      type: object
      properties:
        briefText:
          type: string
      required:
        - briefText
    ParseBriefResponse:
      type: object
      properties:
        clientName:
          type: string
        industry:
          type: string
        proposalContent:
          type: string
      required:
        - clientName
        - industry
        - proposalContent
    GoogleDocsExportResponse:
      type: object
      properties:
        docId:
          type: string
        docUrl:
          type: string
        title:
          type: string
      required:
        - docId
        - docUrl
        - title
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error