# lib/api-zod/src/generated/types/proposal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProposalStatus } from "./proposalStatus";

export interface Proposal {
  id: number;
  clientName: string;
  industry: string;
  status: ProposalStatus;
  briefText: string;
  proposalContent: string;
  googleDocUrl?: string | null;
  createdAt: Date;
  updatedAt: Date;
}