# lib/api-zod/src/generated/types/updateProposalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProposalBodyStatus } from "./updateProposalBodyStatus";

export interface UpdateProposalBody {
  clientName?: string;
  industry?: string;
  proposalContent?: string;
  status?: UpdateProposalBodyStatus;
}